+++ Application/module/person/general.cpp
@@ -71,8 +71,8 @@ void ModulePerson::Update()
 	}
 
 	// move person attached to active camera
-	auto cam = Entity->Get<Camera>(*Global->Get<unsigned int>("camera"));
-	unsigned int id = cam->Person;
+	auto cam = Entity->Get<Camera>(*Global->Get<uint64_t>("camera"));
+	uint64_t id = cam->Person;
 
 	if(id)
 	{
@@ -122,7 +122,7 @@ void ModulePerson::Update()
 void ModulePerson::Listeners()
 {
 	Event->Listen("InputBindJump", [=]{
-		unsigned int id = Entity->Get<Camera>(*Global->Get<unsigned int>("camera"))->Person;
+		uint64_t id = Entity->Get<Camera>(*Global->Get<uint64_t>("camera"))->Person;
 		if(!id) return;
 		auto psn = Entity->Get<Person>(id);
 		auto frm = Entity->Get<Form>(id);
@@ -140,11 +140,11 @@ void ModulePerson::Listeners()
 		// if there is no person at all, add a default one at camera position
 		if(Entity->Get<Person>().size() < 1)
 		{
-			unsigned int id = Entity->New();
+			uint64_t id = Entity->New();
 			auto psn = Entity->Add<Person>(id);
 			auto frm = Entity->Add<Form>(id);
 			
-			vec3 position = Entity->Get<Form>(*Global->Get<unsigned int>("camera"))->Position() - vec3(0, psn->Eyes, 0);
+			vec3 position = Entity->Get<Form>(*Global->Get<uint64_t>("camera"))->Position() - vec3(0, psn->Eyes, 0);
 			psn->Calculate(1.80f);
 			frm->Position(position);
 
@@ -152,17 +152,17 @@ void ModulePerson::Listeners()
 		}
 
 		// if there is no person bound to camera, bound the first one
-		auto cam = Entity->Get<Camera>(*Global->Get<unsigned int>("camera"));
+		auto cam = Entity->Get<Camera>(*Global->Get<uint64_t>("camera"));
 		if(!cam->Person)
 		{
-			unsigned int id = Entity->Get<Person>().begin()->first;
+			uint64_t id = Entity->Get<Person>().begin()->first;
 			cam->Person = id;
 			Debug->Pass("not bound to camera, fallback to first one");
 		}
 
 		// add player position output
 		Entity->Add<Print>(Entity->New())->Text = [=]()-> string{
-			unsigned int id = Entity->Get<Camera>(*Global->Get<unsigned int>("camera"))->Person;
+			uint64_t id = Entity->Get<Camera>(*Global->Get<uint64_t>("camera"))->Person;
 			if(!id) return "";
 			vec3 position = Entity->Get<Form>(id)->Position();
 			float height = Entity->Get<Person>(id)->Height;
