+++ Application/module/person/module.h
@@ -17,15 +17,15 @@ class ModulePerson : public Module
 	void Callbacks();
 
 	// create
-	void Setup(unsigned int Id);
+	void Setup(uint64_t Id);
 
 	// movement
-	void Move(unsigned int Id, glm::vec3 Amount, float Multiplier = 4.3f);
-	void Jump(unsigned int Id, float Multiplier = 4.8f, bool Force = false);
+	void Move(uint64_t Id, glm::vec3 Amount, float Multiplier = 4.3f);
+	void Jump(uint64_t Id, float Multiplier = 4.8f, bool Force = false);
 
 	// checks
-	bool Ground(unsigned int Id);
-	bool Edge(unsigned int Id, glm::vec3 Direction);
+	bool Ground(uint64_t Id);
+	bool Edge(uint64_t Id, glm::vec3 Direction);
 	std::pair<bool, float> RayDown(btVector3 &Position, float Length);
 	bool Ray(btVector3 &From, btVector3 &To, btVector3 &Point = btVector3(), btVector3 &Normal = btVector3());
